+++ docs/changelog.rst
@@ -21,6 +21,23 @@ Fixed
 -----
 -
 
+Deprecated
+----------
+-
+
+v0.7.1 (29 January 2025)
+========================
+Officially drop support for python 3.9.
+
+Added
+-----
+- **setup_ksatver_vegetation**: method to calculate KsatVer_vegetation to account for biologically-enhanced soil structure in KsatVer. PR #313
+- **setup_lulcmaps_from_vector**: method to prepare LULC map and params from a vector input rather than raster. PR #320
+
+Deprecated
+----------
+- Support for python 3.9 (already not supported in previous releases).
+
 v0.7.0 (8 November 2024)
 ========================
 This release adds support to create water demand and allocation related data (available since Wflow.jl version 0.8.0).